added the found_id to the db, and refactored the classes to make the accessor class more manageble
