+++ src/main/java/data/User.java
@@ -1,0 +1,5 @@
+package data;
+
+public class User {
+
+}
